feat: setup the project
